refactor(payload): move block encoder into module
